Create LivingEntity
