+++ core/src/model/ModelFactory.java
@@ -3,7 +3,7 @@ package model;
 import model.enemies.*;
 
 public class ModelFactory {
-    public static Model makeModel(IMapLoader mapLoader) {
+    public static Model makeModel(IArenaLoader mapLoader) {
         Model model = new Model();
         model.initialize(mapLoader);
         model.addEnemy(new Cyclops(100, 100, 1, 1, 1, model.getWorld()));
