+++ core/src/model/PlayerCharacter.java
@@ -2,7 +2,7 @@ package model;
 
 import com.dongbat.jbump.World;
 
-public class PlayerCharacter extends Entity implements IControllable {
+public class PlayerCharacter extends LivingEntity implements IControllable {
 
     public PlayerCharacter(float spawnX, float spawnY, World<Entity> world) {
         super(spawnX, spawnY, 32, 32, 5, 10, world);
