+++ core/src/model/enemies/Enemy.java
@@ -1,9 +1,10 @@
 package model.enemies;
 
+import model.LivingEntity;
 import model.Entity;
 import com.dongbat.jbump.World;
 
-public abstract class Enemy extends Entity {
+public abstract class Enemy extends LivingEntity {
     private double damage;
 
     public Enemy(float x, float y, float height, float width, float speed, float health, float damage, World<Entity> world) {
